+++ app/models/user.rb
@@ -5,4 +5,12 @@ class User < ApplicationRecord
 
   has_one_attached :avatar
   has_many :stores
+
+  attr_accessor :login
+
+  def self.find_for_database_authentication warden_conditions
+    conditions = warden_conditions.dup
+    login = conditions.delete(:login)
+    where(conditions).where(["lower(username) = :value OR lower(email) = :value", {value: login.strip.downcase}]).first
+  end
 end